Introduce a way to cancel tasks on failure
In done_cb, the application can set cancel_remaining_downloads
boolean variable to abort all enqueued tasks. This can avoid useless
requests when we know in advance that remaining tasks will fail.
This is required to pass some LIGADA DASH tests which are very strict on
the number of requests that the app has to send.