Make the ttmlparse handle two output caps
We need to also push on TTML itself to be able to handle
subtitles as in any other subtitle type: frame by frame.
The TTML is like a demuxer, where it receives maybe just
one buffer with all the subtitle information and then
pushes N buffers. In orde to support TTML on OPE correctly
we need to support having N TTML buffers in the output.
Then the application can choose to either handle the TTLM
by itself or render just that specific frame into RGB or
pango.