Split the parse into two elements
Ok, this is tricky. On oneplay-engine we want to use segmented
TTML. That is, one TTML per each subtitle we want to show, to
keep compatiblity with other subtitles. If we do not do this
fakesink will just receive a single buffer with all the TTML
on it and thus an EOS event will be received. Also, because then
the user will need to parse the TTML himself when we already know
how to do it.

If we keep using a single element, decodebin2 (on a playbin2 based)
pipeline will connect the ttmlparse and will finalize the autoplug
because the plugin has two possible src caps. Then when trying to
compose the subtitles with the video it will work, so good so far.
The problem is on oneplay-engine, the fakesink will connect with
ANY caps and thus there will be no output buffers, don't ask me why.

So what we want to achieve is that playbin2 works the same as before
(with pango markup final caps) and for oneplay-engine we want to always
connect our own ttmlsegmentedparse element. So, we add a new field
on the caps to know if the ttml element is the one we expect and we
increase the rank of the ttmlparse. Playbin2 will always connect
ttmlparse and oneplay-engine will always connect ttmlsegmentedparse.
To achieve this on oneplay-engine we need to dynamically increase
the ttmlsegmentedparse and that's all.