Bump version to 0.10.11
On 5-Oct-2017, tag 0.10.10 (created the 13-Mar-2017) was renamed 0.10.11
(Due to the CodeWeavers Incident, where we forgot to create tags in
Nov-2016 and then we created them later (in Oct-2017) and renamed all
subsequent tags).
libfludownloader does not use tags to generate version numbers, so the
hard-coded version 0.10.10 does not match the tag 0.10.11 from Mar-2017
to now. This commit fixes that.