Use input buffer timestamp to sinchronize clear buffer generation
Without this, ttml render generated clear buffers to fill gaps
only when it parsed active spans. If it received a clear buffer,
it could not generate the clear buffer until it received the next
buffer with active spans.

Now, for any input buffer, fully or partially clear, it generates
clear output buffers up to the time specified by a valid timestamp
and duration of the input buffer.

This may simplify automated testing by ensuring that ttmlrender will
produce output for every valid input buffer, empty or not.