Fix force-clear-buffer=false
A previous change to condition buffer generation to
force-clear-buffer resulted in no buffer generation at all
when set to false.

This fix still does not warrantee that no clear buffers are generated,
some may still be generated. In order to ensure that only buffers with
content are generated when set to false, many more changes and
testing would be necessary. Since it needs to be set to true anyway
to work properly, it is the default value, and there are no known
issues setting it to true, further work is not a priority right now.