Fix default cell resolution
The default cell resolution is 32x15 in TTML standard and 40x24 in
EBU-TT standard. Now we initialize it to the proper value based in
the namespaces declared. We are assuming that an xml declaring
the ebu:tt namespace conforms to EBU-TT.

This required added parsing for namespaces and dumping them in
segmentedttml so that render receives it, as this is required in
the render.