Fix line break handling
The previous fix had an undesirable effect, it sometimes added
newlines where not appropiate, since it always did on final span
closures. The parser is currently generating spans on sax characters
without accumulating til span closure.

Now it has been fixed to work as required in all cases. The original
design does not store paragraphs, only spans. SegmentedTtml
regenerated paragraph closures on newlines embedded in the spans,
but because it does not store or generate the paragraph attributes,
it didn't work.

The generated paragraphs need to include at least the "region" attribute, so that the newlines are generated for the right region.
Now we just include all span attributes in the generated paragraph
using the same code used to dump the span attributes. This can be
further optimized by filtering which attributes are included for
paragraphs and spans to avoid unnecessary duplication. At least
"region" must be in the paragraphs and is not necessary in spans.

Now segmentedTtml generates an attribute space="preserve", instructing upstream parsers not to filter out space, since
segmentedTtml already did. It does not change behavior anyway,
it is just a cheap but effective optimization.