Fix wrong span splitting in ttml parsing
We were generating spans directly from sax_characters, but it may
not provide the whole span content, so we were generating partial
spans and incorrectly removing space per each chunk.

Now we accumulate sax_characters in a buffer and generate the spans
properly at node closure. This buffer can be used for any kind of
node, and embedded images was changed to use it too. It is a single
buffer that is reallocated by chucks always larger than needed,
to minimize allocations and reallocations.

The white space removal code has been moved to a function,
improving and substituting similar code in several places. Now it is
used directly on the sax buffer before data is copied, reducing
complexity, allocations and copies.

The segmented parser span and paragraph dump has been simplified and
improved to always generate paragrapg closures on newlines and
avoiding inclussion of the newlines in dumped spans.