Fix cellResolution and fontSize handling
Dump cellResolution attribute.

Fix pango fontSize generation, it must be an integer.

Fix fontSize normalization.

In paragraphs, dump only the region attribute to avoid incorrect
nested attributes normalization. We normalize attributes in place,
losing the ability to discern if the original was relative or
absolute, we cannot handle nested attributes properly. This does not
seem to cause issues with known tests thought, and support for it
would require major changes.