Fix xml document splitter before passing to sax parser
We were looking for root node closure as "</tt>". But xml documents may
not use the tt namespace as default and we will find "</ttns:tt> where ttns
is an arbitrary namspace alias impossible to determine before parsing.

Since what we need is to detect document boundaries in the stream, we fixed
the bug by looking for the document start tag "<?xml", much better and robust.

There is still another problem with this implementation: it may fail if this
tag is split in different buffers. We could rely instead in document start and
end callbacks from the sax parser, we should not need to split documents before
passing to a sax parser. But this further fix would need meaningful additional
investigation and testing and we do not want to invest in it right now.