Fix ttml typefind
The previous fix to handle root node with namespace had a flaw: typefind API
does not allow the implemented optimizations.
This is a new implementation handling root node with namespaces made in the
less efficient way used previously.
The check for "<?xml" was removed because some test files do not have it.