Fix ttml pre-parse to handle root node with namespace
We should rely in the sax parser to notify end of document.
But seems that our xml2 sax parser does not work as expected or we are not
using it in a way that works, it does not start parsing until we tell it
that we received the end of document.

For now, we keep this ugly code to detect documend ends and split if necessary,
but fixed to detect end of documents when the root node is not in the default
namespace.