Fix handling of abortion in data callback
Abortion in progress callback was reported correctly with outcome aborted,
but the less common abortion in the data callback was being incorrectly
reported as task error.

If data arrive after an error, we must not stream it and we were aborting
in the data callback. But the original error must not change, so now we just
ignore the data without aborting. Now the behaviour is correct. We
may consider later an optimization to allow abortion while keeping the http
error that caused it, keeping the same behaviour.

Minor improvements in headers handling.