Fix timeout transition from connect to data with redirections
We were changing the timeout to the shorted one of data when receiving headers.
In case of redirection, the next connection could fail because the wrong
shorter timeout was used. Now we only change the timeout for data for final
status 2xx.

To ease handling of http_status in different places and eliminate the chances
of bugs, a new bool http_status_error has been added.
http_status_ok means a final OK (2xx), http_status_error means a final
error (>400), and these are computed and updated only in one place.