Revert "Allow build with glib version < 2.32"
This reverts commit ae60c16fabfd2a1959494e6f16ab98750dc8d7be [formerly a1768f3408e7a8551a69216937917d23431fb139].

The fix was not enough to make fluc/threads compatible with glib versions older
than 2.32, because support for required functionality such recursive mutexes
and monotonic time was introduced in 2.32.
The fix does not do any harm, it has no effect at all with glib >= 2.32, but is
being removed anyway to keep things clean and avoid any confussion regarding
support for glib < 2.32.


Former-commit-id: 17d09eaca6f0d477605840cbd34f421c132dc9fd