meson: Refactor of symbols visibility logic
There are some cases where we need to fallback to "symbols.map" because the
appropriate compiler/liker flags don't work correctly. The best way to
achieve this is to apply the "hide_symbols_ldflags" only to the plugins but
not to their libraries (ie. libippmerged). This way we avoid adding
duplicated flags and achieve a simpler code.

Issue: OCP_3701