CI/common: delete common and create a python version of git-product-version
Current GHA implementation fails when a workflow is executed after a
release. CI requires the hash of the commit in the version string of
the plugin, but the version is the same as the last release, because no
commits were added after the tag. Error is:

```
Cannot find string PLUGIN_VERSION .*8e672c180, in .fluendo section of the binary
```

To use the correct version to check in the CI, a new
git-product-version.py is used by the check in GHA, by meson, and by
fluendo-cerbero. To use git-product-version.py in the CI,
the checkout was updated to include all the history and the tags.

Without git-product-version, the common submodule is not needed
anymore. Only for generic headers that were copied from the
common subproject:

* libs/common/glib-compat.h
* libs/common/gst-compat.h
* libs/common/gst-demo.h
* libs/common/gst-fluendo.h

NOTE: Using git describe changes version format adding a `g` before the
hash.

Before: 1.0.6-161-1358d5f754
After:  1.0.6-161-g1358d5f75

Issue: OCP-5158 (OCP-4821 and OCP-4825)